refac: abstract personal portfolios comp
