+++ libs/features/src/lib/project/components/organisms/index.ts
@@ -1,2 +1,3 @@
 export * from './project-card/project-card';
 export * from './company-portfolios/company-portfolios';
+export * from './personal-portfolios/personal-portfolios';
