+++ libs/features/src/lib/project/components/organisms/personal-portfolios/personal-portfolios.spec.tsx
@@ -1,0 +1,12 @@
+import { render } from '@testing-library/react';
+
+import PersonalPortfolios from './personal-portfolios';
+
+describe('PersonalPortfolios', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(
+      <PersonalPortfolios />
+    );
+    expect(baseElement).toBeTruthy();
+  });
+});
