+++ libs/features/src/lib/project/components/organisms/personal-portfolios/personal-portfolios.stories.tsx
@@ -1,0 +1,19 @@
+import { Story, Meta } from '@storybook/react';
+import { within } from '@storybook/testing-library';
+import { PersonalPortfolios } from './personal-portfolios';
+
+export default {
+  component: PersonalPortfolios,
+  title: 'organisms/PersonalPortfolios',
+  argTypes: {},
+} as Meta;
+
+const Template: Story = (args) => <PersonalPortfolios {...args} />;
+
+export const Primary = Template.bind({});
+Primary.args = {};
+
+Primary.play = async ({ canvasElement }) => {
+  const canvas = within(canvasElement);
+  // TODO SOMETHING
+};
